[hotfix][python] Remove unnecessary classes such as AbstractRowDataPythonScalarFunctionOperator and AbstractPythonTableFunctionOperator
