+++ flink-table/flink-table-planner/src/main/java/org/apache/flink/table/planner/plan/nodes/exec/common/CommonExecPythonCalc.java
@@ -66,11 +66,11 @@ public abstract class CommonExecPythonCalc extends ExecNodeBase<RowData>
 
     private static final String PYTHON_SCALAR_FUNCTION_OPERATOR_NAME =
             "org.apache.flink.table.runtime.operators.python.scalar."
-                    + "RowDataPythonScalarFunctionOperator";
+                    + "PythonScalarFunctionOperator";
 
     private static final String ARROW_PYTHON_SCALAR_FUNCTION_OPERATOR_NAME =
             "org.apache.flink.table.runtime.operators.python.scalar.arrow."
-                    + "RowDataArrowPythonScalarFunctionOperator";
+                    + "ArrowPythonScalarFunctionOperator";
 
     @JsonProperty(FIELD_NAME_PROJECTION)
     private final List<RexNode> projection;
